+++ app/renderer/common.js
@@ -1,6 +1,6 @@
 /*
 * Vieb - Vim Inspired Electron Browser
-* Copyright (C) 2021-2022 Jelmer van Arnhem
+* Copyright (C) 2021-2023 Jelmer van Arnhem
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
@@ -46,7 +46,7 @@ const currentMode = () => document.body.getAttribute("current-mode") || "normal"
 const getSetting = val => getStored("settings")?.[val]
 
 const getMouseConf = val => {
-    const mouse = getStored("settings")?.mouse
+    const mouse = getSetting("mouse")
     return mouse?.split(",").includes("all") || mouse?.split(",").includes(val)
 }